requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
jinja2==3.1.4
python-multipart==0.0.9
stripe==9.9.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
itsdangerous==2.2.0
qrcode[pil]==7.4.1
